fix pill state reset when navigating tabs
Tab navigation causes remounting of `HomeHeader` component (not
fixable), which resets the  internal state of `useLocation` hook that's used within it.

To fix, this commit moves the usage of the hook outside the rendering of the `HomeHeader` and provides the
necesssary location-based data as props.
